Rename "primitives" to "util"
"util" is a much more common term so this better tells the reader the purpose of the component under this directory.
